Debugged and commented out problems
There were several errors in my last upload, so I've fixed a bunch of those problems. However, it still does not like my FriendshipRepository. I've commented out the code for addFriend and removeFriend for now so that it compiles. I believe the issue lies in my making an explicit constructor for Friendship. I also commented out the forms I added to index, since I'm not sure if they are causing problems with the main page.