Added the add/remove friends features
I've debugged the add and remove friend methods, I think. I've added forms to index.html under the movie shelf sections. It does not seem to work currently, but I do not know if it is because of my code or if it is a separate issue. Currently, when I try to add a friend I get the 500 error. I would appreciate any advice you guys can provide.